requirements.txt
alembic==1.9.2
beautifulsoup4==4.12.0
Flask==2.2.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.3
lxml==4.9.1
pytest==7.1.2
python_dateutil==2.8.2
requests==2.28.1
SQLAlchemy==1.4.39
